\name{clustering}
\alias{clustering}
\title{Clustering Object}
\description{A clustering object is a list with a component 
called \sQuote{clustering} which is an integer vector of length n
where n is the number of elements in a classification.
}
\details{The clustering object is defined simply to allow a cleaner
interface to functions in package \sQuote{cluster}.
}

\references{\url{http://ecology.msu.montana.edu/labdsv/R}}
\author{
    David W. Roberts
    \email{droberts@montana.edu}
}
\keyword{clustering}

