\name{shoshveg}
\docType{data}
\alias{shoshveg}
\title{Vascular Plant Species Cover for the Shoshone National Forest, Wyoming, USA}
\description{Percent cover (in codes) for 368 vascular plants on 150 \eqn{375m^2} sample plots 
stratified across the Shoshone National Forest, Wyoming, USA.  Plots were chosen at
random from a larger dataset.}
\usage{data(shoshveg)}
\format{A data.frame with sample plots as rows and species as columns.  Sample plots
match the \sQuote{shoshsite} dataset.

The cover of all vascular plant species was recorded according to the
following scale: present but $<$ 1\% = 0.1, 1-5\% = 0.5, 5-15\% = 1.0, 15-25\% =
2.0, 25-35\% = 3.0, 35-45\% = 4.0, 45-55\% = 5.0, 55-65\% = 6.0, 65=75\% = 7.0,
75-85\% = 8.0.  No species exhibited greater than 80\% cover in the data set.
}
\note{The data were derived from a multi-year effort by numerous scientists and
field technicians.  The project was directed by Kent Houston, Soil Scientist and
Ecologist, Shoshone National Forest}
\source{Roberts, D.W.  2008.  Statistical Analysis of Multidimensional 
Fuzzy Set Ordinations.  Ecology 89:1246-1260.}
\keyword{datasets}


